backend/requirements.txt
fastapi
uvicorn[standard]
sqlalchemy
psycopg2-binary
python-dotenv
langchain
langgraph
langchain-groq
langchain-openai